display GET values
