+++ package.json
@@ -19,6 +19,7 @@
     "draft-js-export-html": "^1.4.1",
     "draftjs-to-html": "^0.9.1",
     "eslint": "5.6.0",
+    "moment": "^2.29.1",
     "prop-types": "^15.7.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
